refactor: extract excel processing service
