+++ Logibooks.Core.Tests/Controllers/RegistersControllerTests.cs
@@ -56,8 +56,6 @@ public class RegistersControllerTests
     private Role _adminRole;
     private User _logistUser;
     private User _adminUser;
-    private Type _controllerType;
-    private MethodInfo _processExcelMethod;
 #pragma warning restore CS8618
 
     private readonly string testDataDir = Path.Combine(AppContext.BaseDirectory, "test.data");
@@ -100,10 +98,6 @@ public class RegistersControllerTests
         _logger = new LoggerFactory().CreateLogger<RegistersController>();
         _controller = new RegistersController(_mockHttpContextAccessor.Object, _dbContext, _logger);
 
-        _controllerType = typeof(RegistersController);
-        _processExcelMethod = _controllerType.GetMethod("ProcessExcel",
-            BindingFlags.NonPublic | BindingFlags.Instance)
-            ?? throw new InvalidOperationException("ProcessExcel method not found");
 
     }
 
@@ -758,40 +752,6 @@ public class RegistersControllerTests
     }
 
     [Test]
-    public async Task ProcessExcel_ReturnsBadRequest_WhenExcelFileIsEmpty()
-    {
-        SetCurrentUserId(1); // Logist user  
-
-        string testFilePath = Path.Combine(testDataDir, "Register_Empty.xlsx");
-        byte[] excelContent;
-
-        try
-        {
-            excelContent = File.ReadAllBytes(testFilePath);
-        }
-        catch (Exception ex)
-        {
-            Assert.Fail($"Test file not found at {testFilePath}: {ex.Message}");
-            return;
-        }
-
-        var mockFile = CreateMockFile("Register_Empty.xlsx", "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet", excelContent);
-        var result = await _controller.UploadRegister(mockFile.Object);
-
-        Assert.That(result, Is.TypeOf<ObjectResult>());
-        var objResult = result as ObjectResult;
-        Assert.That(objResult!.StatusCode, Is.EqualTo(StatusCodes.Status400BadRequest));
-
-        var errorMessage = objResult.Value as ErrMessage;
-        Assert.That(errorMessage, Is.Not.Null);
-        Assert.That(errorMessage!.Msg, Does.Contain("Пустой файл реестра"));
-
-        // Verify no register was created in the database
-        var registersCount = await _dbContext.Registers.CountAsync();
-        Assert.That(registersCount, Is.EqualTo(0), "No register should be created for an empty Excel file");
-    }
-
-    [Test]
     public async Task UploadRegister_ReturnsBadRequest_WhenZipFileWithoutExcel()
     {
         // Arrange
@@ -873,45 +833,6 @@ public class RegistersControllerTests
         Assert.That(registersCount, Is.EqualTo(0), "No register should be created for unsupported file types");
     }
 
-    private async Task<IActionResult> InvokeProcessExcel(byte[] content, string fileName, string mappingFile = "register_mapping.yaml")
-    {
-        return await (Task<IActionResult>)_processExcelMethod.Invoke(
-            _controller,
-            [content, fileName, mappingFile])!;
-    }
-
-    [Test]
-    public async Task ProcessExcel_Returns500Error_WhenMappingFileNotFound()
-    {
-        // Arrange
-        // Create a sample Excel file content
-        byte[] excelContent = [0x50, 0x4B, 0x03, 0x04]; // Just some dummy content
-
-        // Make sure mapping directory exists but use a non-existent mapping file name
-        string mappingDir = Path.Combine(AppContext.BaseDirectory, "mapping");
-        Directory.CreateDirectory(mappingDir);
-        string nonExistentMappingFile = "non_existent_mapping.yaml";
-        string mappingPath = Path.Combine(mappingDir, nonExistentMappingFile);
-
-        // Delete the mapping file if it somehow exists
-        if (File.Exists(mappingPath))
-        {
-            File.Delete(mappingPath);
-        }
-
-        // Act
-        var result = await InvokeProcessExcel(excelContent, "test.xlsx", nonExistentMappingFile);
-
-        // Assert
-        Assert.That(result, Is.TypeOf<ObjectResult>());
-        var objResult = result as ObjectResult;
-        Assert.That(objResult!.StatusCode, Is.EqualTo(StatusCodes.Status500InternalServerError));
-
-        var errMessage = objResult.Value as ErrMessage;
-        Assert.That(errMessage, Is.Not.Null);
-        Assert.That(errMessage!.Msg, Does.Contain("Не найдена спецификация файла реестра"));
-        Assert.That(errMessage.Msg, Does.Contain(mappingPath));
-    }
 
     [Test]
     public async Task DeleteRegister_DeletesRegisterAndOrders_WhenUserIsLogist()
@@ -974,120 +895,5 @@ public class RegistersControllerTests
             .Returns(Task.CompletedTask);
 
         return mockFile;
-    }
-}
-
-[TestFixture]
-public class ConvertValueToPropertyTypeTests
-{
-#pragma warning disable CS8618
-    private RegistersController _controller;
-#pragma warning restore CS8618
-
-    [SetUp]
-    public void Setup()
-    {
-        // Create a minimal DbContextOptions for AppDbContext
-        var options = new DbContextOptionsBuilder<AppDbContext>()
-            .UseInMemoryDatabase($"test_db_{Guid.NewGuid()}")
-            .Options;
-
-        // Pass the valid options to the AppDbContext constructor
-        var dbContext = new AppDbContext(options);
-
-        // Create the controller instance with the valid AppDbContext
-        _controller = (RegistersController)Activator.CreateInstance(
-            typeof(RegistersController),
-            [new Mock<IHttpContextAccessor>().Object, dbContext, new Mock<ILogger<RegistersController>>().Object]
-        )!;
-    }
-
-    [TestCase("42", typeof(int), 42)]
-    [TestCase("notanint", typeof(int), 0)]
-    [TestCase("3.14", typeof(double), 3.14)]
-    [TestCase("2,71", typeof(double), 2.71)] // comma as decimal separator
-    [TestCase("notadouble", typeof(double), 0.0)]
-    [TestCase("123.45", typeof(decimal), 123.45)]
-    [TestCase("67,89", typeof(decimal), 67.89)]
-    [TestCase("notadecimal", typeof(decimal), 0.0)]
-    [TestCase("true", typeof(bool), true)]
-    [TestCase("false", typeof(bool), false)]
-    [TestCase("1", typeof(bool), true)]
-    [TestCase("0", typeof(bool), false)]
-    [TestCase("yeS", typeof(bool), true)]
-    [TestCase("no", typeof(bool), false)]
-    [TestCase("Да", typeof(bool), true)]
-    [TestCase("нет", typeof(bool), false)]
-    [TestCase("", typeof(bool), false)]
-    [TestCase("notabool", typeof(bool), false)]
-    [TestCase("2024-06-28", typeof(DateTime), "2024-06-28")]
-    [TestCase("notadate", typeof(DateTime), "0001-01-01")]
-    [TestCase("2024-06-28", typeof(DateOnly), "2024-06-28")]
-    [TestCase("2024-06-28T13:00:12", typeof(DateOnly), "2024-06-28")]
-    [TestCase("notadate", typeof(DateOnly), "0001-01-01")]
-    [TestCase("hello", typeof(string), "hello")]
-    [TestCase("", typeof(string), "")]
-    [TestCase(null, typeof(string), "")]
-
-    public void ConvertValueToPropertyType_PrimitiveTypes_Works(string? input, Type type, object expected)
-    {
-        var result = _controller.GetType()
-            .GetMethod("ConvertValueToPropertyType", BindingFlags.NonPublic | BindingFlags.Instance)!
-            .Invoke(_controller, new object?[] { input, type, "TestProp" });
-
-        if (type == typeof(DateTime))
-        {
-            var expectedDate = DateTime.TryParse(expected.ToString(), out var dt) ? dt : default;
-            Assert.That(result, Is.EqualTo(expectedDate));
-        }
-        else if (type == typeof(DateOnly))
-        {
-            var expectedDate = DateOnly.TryParse(expected.ToString(), out var d) ? d : default;
-            Assert.That(result, Is.EqualTo(expectedDate));
-        }
-        else
-        {
-            Assert.That(result, Is.EqualTo(expected));
-        }
-    }
-
-    [Test]
-    public void ConvertValueToPropertyType_NullableInt_ReturnsNullOnNull()
-    {
-        var type = typeof(int?);
-        var result = _controller.GetType()
-            .GetMethod("ConvertValueToPropertyType", BindingFlags.NonPublic | BindingFlags.Instance)!
-            .Invoke(_controller, new object?[] { null, type, "TestProp" });
-        Assert.That(result, Is.Null);
-    }
-
-    [Test]
-    public void ConvertValueToPropertyType_NullableDouble_ReturnsNullOnEmpty()
-    {
-        var type = typeof(double?);
-        var result = _controller.GetType()
-            .GetMethod("ConvertValueToPropertyType", BindingFlags.NonPublic | BindingFlags.Instance)!
-            .Invoke(_controller, new object?[] { "", type, "TestProp" });
-        Assert.That(result, Is.Null);
-    }
-
-    [Test]
-    public void ConvertValueToPropertyType_UnknownType_UsesChangeType()
-    {
-        var type = typeof(long);
-        var result = _controller.GetType()
-            .GetMethod("ConvertValueToPropertyType", BindingFlags.NonPublic | BindingFlags.Instance)!
-            .Invoke(_controller, new object?[] { "123456789", type, "TestProp" });
-        Assert.That(result, Is.EqualTo(123456789L));
-    }
-
-    [Test]
-    public void ConvertValueToPropertyType_UnknownType_ReturnsDefaultOnError()
-    {
-        var type = typeof(Guid);
-        var result = _controller.GetType()
-            .GetMethod("ConvertValueToPropertyType", BindingFlags.NonPublic | BindingFlags.Instance)!
-            .Invoke(_controller, new object?[] { "notaguid", type, "TestProp" });
-        Assert.That(result, Is.EqualTo(Guid.Empty));
     }
 }
